+++ britishcycling_clubs/__init__.py
@@ -1,17 +1,21 @@
 """Module with functions to retrieve information about a club."""
 
 from britishcycling_clubs.manager import (
+    ManagerMemberCounts,
     club_manager_url_via_login,
     get_manager_member_counts,
 )
 from britishcycling_clubs.profile import (
+    ProfileInfo,
     club_profile_url,
     get_profile_info,
 )
 
 __all__ = [
-    "club_manager_url_via_login",
-    "get_manager_member_counts",
     "club_profile_url",
     "get_profile_info",
+    "ProfileInfo",
+    "club_manager_url_via_login",
+    "get_manager_member_counts",
+    "ManagerMemberCounts",
 ]